+++ R/scale.R
@@ -26,7 +26,7 @@ scale_range <- function(x, to = c(0, 1), from = range(x, finite = TRUE)) {
 #' @param na The color to return for `NA` values.
 #' @param ... Currently not used.
 #' @return A [`character`] vector of colors.
-#' @family palette
+#' @family palettes
 #' @name palette_color
 #' @rdname palette_color
 NULL
@@ -72,7 +72,7 @@ palette_color_discrete <- function(x, palette = NULL, domain = NULL,
   }
   if (is.function(palette)) {
     palette <- palette(n, ...)
-  } else {
+  } else if (length(palette) < n) {
     palette <- grDevices::colorRampPalette(palette, ...)(n)
   }
   col <- palette[x]
@@ -88,4 +88,55 @@ make_levels <- function(x, levels = NULL, ordered = FALSE) {
   else if (ordered) levels <- unique(x)
   else levels <- sort(unique(x))
   levels
+}
+
+#' Symbol Mapping
+#'
+#' @param x A vector of categorical values.
+#' @param palette A vector of symbols.
+#' @param domain A vector of categorical data.
+#' @param ordered A [`logical`] scalar: should the levels be treated as already
+#'  in the correct order?
+#' @param ... Currently not used.
+#' @return A vector of symbols.
+#' @family palettes
+#' @export
+palette_shape <- function(x, palette = NULL, domain = NULL,
+                          ordered = FALSE, ...) {
+  domain <- make_levels(x, levels = domain, ordered = ordered)
+  x <- match(as.character(x), domain)
+
+  if (is.null(palette)) {
+    n <- length(domain)
+    if (n > 6) {
+      warning("Consider specifying symbols manually: ",
+              "more than 6 becomes difficult to discriminate.", call. = FALSE)
+    }
+    palette <- c(16, 17, 15, 3, 7, 8)[seq_len(n)]
+  }
+  symb <- palette[x]
+  symb
+}
+
+#' Symbol Size Mapping
+#'
+#' @param x A vector of categorical values.
+#' @param palette A [`numeric`] range.
+#' @param domain A vector of categorical data.
+#' @param ... Currently not used.
+#' @return
+#'  A [`numeric`] vector giving the amount by which plotting text and symbols
+#'  should be magnified relative to the default (see [`par("cex")`][graphics::par()]).
+#' @family palettes
+#' @export
+palette_size <- function(x, palette = NULL, domain = NULL, ...) {
+  if (!is.numeric(x)) {
+    warning("Discrete value supplied to continuous scale.", call. = FALSE)
+    return(rep(1, length(x)))
+  }
+  rng <- range(x, na.rm = TRUE)
+  pal <- if (!is.null(palette)) range(palette, na.rm = TRUE) else rng / max(x, na.rm = TRUE)
+  dom <- if (!is.null(domain)) range(domain, na.rm = TRUE) else rng
+  x <- scale_range(x, from = dom, to = pal)
+  x
 }