+++ NEWS.md
@@ -1,6 +1,7 @@
 # arkhe 1.4.0.9000
 ## New classes and methods
 * Add `palette_color_continuous()` and `palette_color_discrete()` for color mapping.
+* Add `palette_shape()` and `palette_size()` for symbol shape and size mapping.
 
 ## Enhancements
 * `compact()` and `remove_*()` gained a new `verbose` argument to report extra information on progress.